requirements.txt
fastapi==0.111.0              
uvicorn[standard]==0.30.0  
SQLAlchemy==2.0.30            
psycopg2-binary==2.9.9        
geoalchemy2==0.15.2           
shapely==2.0.4                 
pydantic==2.7.2                
python-dotenv==1.0.1   
alembic==1.13.2 
python-multipart==0.0.9
